requirements.txt
flask
websockets
python-dotenv